ln /root/.meteor

+++ build.sh
@@ -7,8 +7,10 @@ echo "Building the app..."
 cd /src
 
 if [ -e /meteor_build_cache/meteor ]; then
-  cp -r /meteor_build_cache/* /root/.meteor/
+  rm -rf /root/.meteor
+  ln -s /meteor_build_cache /root/.meteor /root/.meteor
 fi
+
 set +e # Allow the next command to fail
 meteor build --directory ${APP_DIR}
 if [ $? -ne 0  ]; then
@@ -19,7 +21,9 @@ if [ $? -ne 0  ]; then
   tar xf bundle.tar.gz -C ${APP_DIR}
 fi
 set -e
-mv -f /root/.meteor/* /meteor_build_cache/
+if [ ! -e /meteor_build_cache/meteor ]; then
+  mv -f /root/.meteor/* /meteor_build_cache/
+fi
 # Install NPM modules
 APP_DIR=${APP_DIR}/bundle
 if [ -d ${APP_DIR}/programs/server/node_modules ]; then
